+++ changelog.md
@@ -47,6 +47,7 @@
     - `Logger`: The Logger class, which can be used to create a new logging category. Any created Logger instance will automatically share its logs with BYTM's internal logging system, which can be downloaded using a GM menu command.
     - `getSelector()`: Returns a remotely fetched CSS selector for the given category and name. Refer to `assets/data.json` in BYTM's source code for a list of all values.
     - `sanitizeUnicode()`: Function that replaces all sorts of wacky Unicode characters with their ASCII counterparts if possible. This function is also used by `sanitizeArtists()` and `sanitizeSong()`.
+    - `preventEvents()`: Prevents the given events' listeners from being called and bubbling, for the given container element, when a predicate function matches.
   - Added new properties to the object returned by `BYTM.getInternals()`:
     - `globservers` - Object of all `SelectorObserver` instances used by BYTM.
     - `getSerializerStores()` - Returns all `DataStore` instances that contain user-configured data.